Modify _print.c

+++ p_char.c
@@ -6,7 +6,7 @@
  * Return: number of byte printed
 */
 
-ssize_t p_char(va_list args)
+int p_char(va_list args)
 {
 	char c = va_arg(args, int);
